extract inner class
and inject as dependency
