+++ bukkit/src/main/java/io/github/wysohn/triggerreactor/bukkit/modules/BukkitDriverModule.java
@@ -37,6 +37,7 @@ public class BukkitDriverModule extends AbstractModule {
         bind(IPluginManagement.class).to(BukkitPluginManagement.class);
         bind(TaskSupervisor.class).to(BukkitTaskSupervisor.class);
 
+        bind(BungeeCordHelper.class);
         bind(MysqlSupport.class);
     }
 }
